requirements.txt
flask==3.0.3
flask_sqlalchemy==3.0.5
requests==2.31.0
psycopg2-binary==2.9.9
